MANTA-527: Have mola take care of GCing mako cruft (adding no verification option)

+++ bin/cruft_verify_and_link.js
@@ -58,12 +58,19 @@ function parseOptions() {
         var option;
         var jsonOpts = fs.readFileSync(MOLA_CONFIG);
         var opts = JSON.parse(jsonOpts);
-        var parser = new getopt.BasicParser('L',
+        var parser = new getopt.BasicParser('hLX',
                                             process.argv);
         while ((option = parser.getopt()) !== undefined && !option.error) {
                 switch (option.option) {
+                case 'h':
+                        usage();
+                        process.exit(1);
+                        break; //Makes jsl happy
                 case 'L':
                         opts.noLink = true;
+                        break;
+                case 'X':
+                        opts.noVerify = true;
                         break;
                 default:
                         usage('Unknown option: ' + option.option);
@@ -80,7 +87,11 @@ function usage(msg) {
                 console.error(msg);
         }
         var str  = 'usage: ' + path.basename(process.argv[1]);
-        str += ' [-L verify but don\'t link]';
+        str += ' [-L don\'t link]';
+        str += ' [-X don\'t verify]\n';
+        str += '\n';
+        str += 'Don\'t use -X unless you\'ve first run with a -L and are\n';
+        str += '*sure* of successful verification.';
         console.error(str);
         process.exit(1);
 }
@@ -337,6 +348,13 @@ function verifyAndLinkFile(opts, cb) {
         vasync.pipeline({
                 'funcs': [
                         function getToFile(_, subcb) {
+                                if (opts.noVerify) {
+                                        LOG.info({
+                                                'path': cruft,
+                                                'file': file
+                                        }, 'NOT verifying file');
+                                        return (subcb());
+                                }
                                 LOG.info({
                                         'path': cruft,
                                         'file': file
@@ -348,6 +366,9 @@ function verifyAndLinkFile(opts, cb) {
                                 }, subcb);
                         },
                         function checkFile(_, subcb) {
+                                if (opts.noVerify) {
+                                        return (subcb());
+                                }
                                 LOG.info({
                                         'path': cruft,
                                         'file': file
@@ -367,6 +388,9 @@ function verifyAndLinkFile(opts, cb) {
                                 });
                         },
                         function rmTmpFile(_, subcb) {
+                                if (opts.noVerify) {
+                                        return (subcb());
+                                }
                                 fs.unlink(file, subcb);
                         },
                         function linkFiles(_, subcb) {
